cargo: Add enum-iterator to prevent 1.2.0 from being used.
`enum-iterator` is a dependency for `vergen`, but because `v1.2.0` is causing
the `error[E0658]`, we are restricting its version by defining it as `~1.1.3`
here.

To give more context, until we've added tracing, vergen v1.1.3 had been chosen
by cargo. With tracing, if we use any version for vergen equal or greater than
v7.0.0, cargo jumps enum-iterator from v0.8.1 to v1.2.0, which will cause an
error when building in the CI:

error[E0658]: use of unstable library feature 'array_from_fn'
   --> /root/.cargo/registry/src/github.com-1285ae84e5963aae/enum-iterator-1.2.0/src/lib.rs:554:18
    |
554 |             Some(core::array::from_fn(|_| unreachable!()))
    |                  ^^^^^^^^^^^^^^^^^^^^
    |
    = note: see issue #89379 <https://github.com/rust-lang/rust/issues/89379> for more information

By adding `enum-iterator = "~1.1.3"`, we are giving v1.1.3 preference
over the latest version (v1.2.0), so the build is fixed.