src: lib: stream: pipeline: Improve freeze detection strategy
The previous strategy was not enough to deal with lower FPS streams,
triggering stream recreation when not needed, disrupting the reliability
instead of helping with it.

This new strategy uses the nominal FPS coming from the stream information
to calculate the frame interval, and when not available (like in redirect UDP),
it learns the expected FPS dynamically, triggering stream recreation upon large
deviations.